refactor: split useAttachmentUploads.test.ts
